Profile API demo

+++ src/main/java/com/smileflower/santa/flag/repository/FlagRepository.java
@@ -1,5 +1,5 @@
 package com.smileflower.santa.flag.repository;
 
 public interface FlagRepository {
-    int updateImageUrlByIdx(Long userIdx, Long mountainIdx, String filename);
+    int updateImageUrlByIdx(int userIdx, Long mountainIdx, String filename);
 }
