+++ src/main/java/com/smileflower/santa/profile/repository/ProfileRepository.java
@@ -13,31 +13,31 @@ import java.util.Optional;
 public interface ProfileRepository {
     //단일 유저 조회
     Optional<Profile> findByEmail(String email);
-    Optional<Profile> findByIdx(Long userIdx);
+    Optional<Profile> findByIdx(int userIdx);
     //닉네임 조회
-    String findNameByIdx(Long userIdx);
+    String findNameByIdx(int userIdx);
     //게시물(정복사진,그냥사진) 조회
-    List<FlagResponse> findFlagsByIdx(Long userIdx);
-    List<Picture> findPicturesByIdx(Long userIdx);
+    List<FlagResponse> findFlagsByIdx(int userIdx);
+    List<Picture> findPicturesByIdx(int userIdx);
 
-    List<FlagsForMapResponse> findFlagsForMapByIdx(Long userIdx);
+    List<FlagsForMapResponse> findFlagsForMapByIdx(int userIdx);
     //정복 횟수
-    int findFlagCountByIdx(Long userIdx);
+    int findFlagCountByIdx(int userIdx);
     //사진 올리기
-    int createPicture(Long userIdx,String imageUrl);
+    int createPicture(int userIdx,String imageUrl);
     //게시물 삭제
     boolean deleteFlagByIdx(Long flagIdx);
     boolean deletePictureByIdx(Long flagIdx);
     //정복사진 신고
-    Long report(Long flagIdx, Long userIdx);
+    Long report(Long flagIdx, int userIdx);
     //정복사진 신고 카운트
     int reportCountByIdx(Long flagIdx);
 
     int updateImageUrlByEmail(String email,String filename);
 
-    int updateImageUrlByIdx(Long userIdx, String filename);
+    int updateImageUrlByIdx(int userIdx, String filename);
 
     int deleteImageUrlByEmail(String email);
 
-    int deleteImageUrlByIdx(Long userIdx);
+    int deleteImageUrlByIdx(int userIdx);
 }
